+++ src/components/BootstrapDialog/BootstrapDialog.jsx
@@ -1,8 +1,5 @@
-import CloseIcon from '@mui/icons-material/Close';
-import { Dialog, DialogTitle, IconButton } from '@mui/material';
+import { Dialog } from '@mui/material';
 import { styled } from '@mui/material/styles';
-
-import PropTypes from 'prop-types';
 
 const BootstrapDialog = styled(Dialog)(({ theme }) => ({
   '& .MuiDialogContent-root': {
@@ -13,32 +10,4 @@ const BootstrapDialog = styled(Dialog)(({ theme }) => ({
   },
 }));
 
-const BootstrapDialogTitle = (props) => {
-  const { children, onClose, ...other } = props;
-
-  return (
-    <DialogTitle sx={{ m: 0, p: 2 }} {...other}>
-      {children}
-      {onClose ? (
-        <IconButton
-          aria-label='close'
-          onClick={onClose}
-          sx={{
-            position: 'absolute',
-            right: 8,
-            top: 8,
-            color: (theme) => theme.palette.grey[500],
-          }}>
-          <CloseIcon />
-        </IconButton>
-      ) : null}
-    </DialogTitle>
-  );
-};
-
-BootstrapDialogTitle.propTypes = {
-  children: PropTypes.node,
-  onClose: PropTypes.func.isRequired,
-};
-
-export { BootstrapDialog, BootstrapDialogTitle };
+export default BootstrapDialog;
