+++ src/components/BootstrapDialog/BootstrapDialogTitle.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+
+import CloseIcon from '@mui/icons-material/Close';
+import { DialogTitle, IconButton } from '@mui/material';
+
+const BootstrapDialogTitle = (props) => {
+  const { children, onClose, ...other } = props;
+
+  return (
+    <DialogTitle sx={{ m: 0, p: 2 }} {...other}>
+      {children}
+      {onClose ? (
+        <IconButton
+          aria-label='close'
+          onClick={onClose}
+          sx={{
+            position: 'absolute',
+            right: 8,
+            top: 8,
+            color: (theme) => theme.palette.grey[500],
+          }}>
+          <CloseIcon />
+        </IconButton>
+      ) : null}
+    </DialogTitle>
+  );
+};
+
+export default BootstrapDialogTitle;
