+++ src/components/index.js
@@ -3,3 +3,5 @@ export { default as Sidebar } from './Sidebar/Sidebar';
 export { default as Bottombar } from './Bottombar/Bottombar';
 export { default as Navbar } from './Navbar/Navbar';
 export { default as PrivateRoute } from './PrivateRoute/PrivateRoute';
+export { default as BootstrapDialog } from './BootstrapDialog/BootstrapDialog';
+export { default as BootstrapDialogTitle } from './BootstrapDialog/BootstrapDialogTitle';
