+++ src/pages/OnboardingPage/components/JoinWorkspaceModal.jsx
@@ -4,10 +4,7 @@ import GroupAddIcon from '@mui/icons-material/GroupAdd';
 import { Box, Button, DialogActions, DialogContent, TextField, Typography } from '@mui/material';
 import { grey, lightBlue } from '@mui/material/colors';
 
-import {
-  BootstrapDialog,
-  BootstrapDialogTitle,
-} from '../../../components/BootstrapDialog/BootstrapDialog';
+import { BootstrapDialog, BootstrapDialogTitle } from '../../../components';
 import { flexCenter, flexCenterColumn } from '../../../styles/commonObjectStyles';
 
 const JoinWorkspaceModal = () => {
